Update screen.css
The bookcreator plugin expects the content of the page to be inside #dokuwiki__content tag instead of #content to load the name of the pages into the "Create pdf book" page after adding pages into the book. This pull request fix the CSS files and main.php file to reflect thoses changes.